Use casts to bridge mismatches in integer-type sizes and signedness
The compiler gets nervous when we (for example) pass in a size_t as an int
parameter, or assign an int to a char, or assign -1 to an unsigned type. By
adding appropriate casts, we inform the compiler that, yes, we know what
we're doing.

This change also upgrades a va_arg() type from char to int, as char is
always promoted to int when passed through '...'.

+++ export.cpp
@@ -503,7 +503,7 @@ void VectorFileWriter::Output(SBezierLoopSetSet *sblss, SMesh *sm) {
             b = sbl->l.First();
             if(!b || !Style::Exportable(b->auxA)) continue;
 
-            hStyle hs = { b->auxA };
+            hStyle hs = { (DWORD)b->auxA };
             Style *stl = Style::Get(hs);
             double lineWidth  = Style::WidthMm(b->auxA)*s;
             DWORD strokeRgb = Style::Color(hs, true);
